backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
rasterio==1.4.3
gunicorn==23.0.0
python-dotenv==1.0.1